Create Atm Interface using Python
The ATMs in our cities are built on python, as we have all seen them. It is a console-based application with five different classes. In order to use the system, the user must enter his or her user ID and pin when it starts. Once the details are entered successfully, ATM functionality is unlocked. As a result of the project, the following operations can be performed:

1.Transactions History

2.Withdraw

3.Deposit

4.Transfer

5.Quit